requirements.txt
feedparser
requests
google-generativeai
python-dotenv
